requirements.txt
jinja2
werkzeug
pyyaml
mwclient
mwtemplates
oursql
beautifulsoup4
